+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/case-flag-summary-list/case-flag-summary-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { FlagDetailDisplay } from '../../domain';
+import { FlagDetail, FlagDetailDisplay } from '../../domain';
 import { CaseFlagFieldState, CaseFlagSummaryListDisplayMode } from '../../enums';
 
 @Component({
@@ -14,6 +14,10 @@ export class CaseFlagSummaryListComponent implements OnInit {
   public flagDescription: string;
   public flagComments: string;
   public flagStatus: string;
+  public flagDescriptionWelsh: string;
+  public flagCommentsWelsh: string;
+  public otherDescription: string;
+  public otherDescriptionWelsh: string;
   public displayMode = CaseFlagSummaryListDisplayMode;
   public addUpdateFlagHeaderText: string;
   public caseFlagFieldState = CaseFlagFieldState;
@@ -24,13 +28,19 @@ export class CaseFlagSummaryListComponent implements OnInit {
   public ngOnInit(): void {
     if (this.flagForSummaryDisplay) {
       const flagDetail = this.flagForSummaryDisplay.flagDetail;
-      this.flagDescription = `${flagDetail.name}${flagDetail.otherDescription
-        ? ` - ${flagDetail.otherDescription}`
-        : ''}${flagDetail.subTypeValue ? ` - ${flagDetail.subTypeValue}` : ''}`;
+      this.flagDescription = this.getFlagDescription(flagDetail);
+      this.flagDescriptionWelsh = flagDetail.otherDescription_cy;
       this.flagComments = flagDetail.flagComment;
+      this.flagCommentsWelsh = flagDetail.flagComment_cy;
       this.flagStatus = flagDetail.status;
       this.addUpdateFlagHeaderText =
         this.summaryListDisplayMode === CaseFlagSummaryListDisplayMode.MANAGE ? this.updateFlagHeaderText : this.addFlagHeaderText;
     }
+  }
+
+  private getFlagDescription(flagDetail: FlagDetail): string {
+    const otherDescription = flagDetail.otherDescription ? ` - ${flagDetail.otherDescription}` : '';
+    const subTypeValue = flagDetail.subTypeValue ? ` - ${flagDetail.subTypeValue}` : '';
+    return `${flagDetail.name}${otherDescription}${subTypeValue}`;
   }
 }