+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/enums/case-flag-form-fields.enum.ts
@@ -1,10 +1,10 @@
 export enum CaseFlagFormFields {
   FLAG_TYPE = 'flagType',
   COMMENTS = 'flagComment',
-  COMMENTS_WELSH = 'flagCommentWelsh',
+  COMMENTS_WELSH = 'flagComment_cy',
   OTHER_FLAG_DESCRIPTION = 'otherDescription',
-  OTHER_FLAG_DESCRIPTION_WELSH = 'otherDescriptionWelsh',
-  STATUS = 'flagStatus',
+  OTHER_FLAG_DESCRIPTION_WELSH = 'otherDescription_cy',
+  STATUS = 'status',
   STATUS_CHANGE_REASON = 'flagStatusReasonChange',
   IS_WELSH_TRANSLATION_NEEDED = 'flagIsWelshTranslationNeeded'
 }